Minor QoL and prettification changes
Calendar Widget click opens thunderbird

Clicking on the calendar widget now opens thunderbird if it wasn't already open.
In future I want to make it so it takes me to the right place in thunderbird

Added day to date widget.

The day is now shown as part of the calendar widget. I was finding I had no idea
what day it was

Added extra space in volume widget

When volume at full, it looked very squished. A space has been added between the
emoji and the percentage number